api_server_django/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
Django==3.2.6
django-cors-headers==3.7.0
django-seed==0.2.2
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.3
Faker==8.11.0
filelock==3.0.12
idna==3.2
importlib-metadata==4.6.3
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
platformdirs==2.2.0
protobuf==3.17.3
PyMySQL==1.0.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
urllib3==1.26.6
validate-email==1.3
virtualenv==20.7.0
zipp==3.5.0
